+++ tests/OpenSleigh.Core.Tests/Sagas/EmptySaga.cs
@@ -1,0 +1,4 @@
+﻿namespace OpenSleigh.Core.Tests.Sagas
+{
+    public class EmptySaga : Saga<DummySagaState> { }
+}
