+++ tests/OpenSleigh.Core.Tests/Unit/DefaultSagaFactoryTests.cs
@@ -3,6 +3,7 @@ using FluentAssertions;
 using Microsoft.Extensions.DependencyInjection;
 using NSubstitute;
 using OpenSleigh.Core.Messaging;
+using OpenSleigh.Core.Tests.Sagas;
 using Xunit;
 
 namespace OpenSleigh.Core.Tests.Unit
