+++ tests/OpenSleigh.Core.Tests/Unit/DefaultSagaStateFactoryTests.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using OpenSleigh.Core.Exceptions;
+using OpenSleigh.Core.Tests.Sagas;
 using Xunit;
 
 namespace OpenSleigh.Core.Tests.Unit
