+++ tests/OpenSleigh.Core.Tests/Unit/DependencyInjection/LambdaSagaStateFactoryTests.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using FluentAssertions;
+using OpenSleigh.Core.Tests.Sagas;
 using Xunit;
 
 namespace OpenSleigh.Core.Tests.Unit.DependencyInjection
