+++ tests/OpenSleigh.Core.Tests/Unit/Messaging/DefaultMessageBusTests.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using NSubstitute;
 using OpenSleigh.Core.Messaging;
 using OpenSleigh.Core.Persistence;
+using OpenSleigh.Core.Tests.Sagas;
 using Xunit;
 
 namespace OpenSleigh.Core.Tests.Unit.Messaging
