+++ tests/OpenSleigh.Core.Tests/Unit/SagaStateServiceTests.cs
@@ -6,6 +6,7 @@ using NSubstitute;
 using OpenSleigh.Core.Exceptions;
 using OpenSleigh.Core.Messaging;
 using OpenSleigh.Core.Persistence;
+using OpenSleigh.Core.Tests.Sagas;
 using Xunit;
 
 namespace OpenSleigh.Core.Tests.Unit
