+++ tests/OpenSleigh.Core.Tests/Unit/SagaStateTests.cs
@@ -1,6 +1,7 @@
 ﻿using FluentAssertions;
 using System;
 using OpenSleigh.Core.Messaging;
+using OpenSleigh.Core.Tests.Sagas;
 using Xunit;
 
 namespace OpenSleigh.Core.Tests.Unit
