+++ tests/OpenSleigh.Core.Tests/Unit/SagaTests.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using FluentAssertions;
 using OpenSleigh.Core.Messaging;
+using OpenSleigh.Core.Tests.Sagas;
 using Xunit;
 
 namespace OpenSleigh.Core.Tests.Unit
